requirements.txt
# Core essencial
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Banco vetorial
faiss-cpu>=1.11.0
sentence-transformers>=4.1.0

# Agentes IA
crewai>=0.95.0
crewai-tools>=0.25.0
google-generativeai>=0.3.0

# Interface e API
streamlit>=1.30.0
fastapi>=0.100.0
uvicorn>=0.18.0

# Utilitários
plotly>=5.15.0
tqdm>=4.65.0